Exception handling

+++ src/main/java/com/devsuperior/dscommerce/service/ProductService.java
@@ -1,16 +1,20 @@
-package com.devsuperior.dscommerce;
-
-import java.util.List;
+package com.devsuperior.dscommerce.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.service.exceptions.DatabaseException;
+import com.devsuperior.dscommerce.service.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -25,7 +29,8 @@ public class ProductService {
 		ProductDTO dto = new ProductDTO(product);
 		return dto;*/
 		
-		Product product = repository.findById(id).get();
+		Product product = repository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProductDTO(product);
 	}
 	
@@ -45,16 +50,30 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDToToEntity(dto, entity);		
-		entity = repository.save(entity);		
-		return new ProductDTO(entity);
+		
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDToToEntity(dto, entity);		
+			entity = repository.save(entity);		
+			return new ProductDTO(entity);
+		}
+		catch(EntityNotFoundException e){
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
-		
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
 
 	private void copyDToToEntity(ProductDTO dto, Product entity) {
